// src/types/datasources/mysql.types.ts
export enum MySQLColumnType {
	INT = 'int',
	TINYINT = 'tinyint',
	SMALLINT = 'smallint',
	MEDIUMINT = 'mediumint',
	BIGINT = 'bigint',
	FLOAT = 'float',
	DOUBLE = 'double',
	DECIMAL = 'decimal',
	NUMERIC = 'numeric',
	REAL = 'real',
	TIMESTAMP = 'timestamp',
	CHAR = 'char',
	VARCHAR = 'varchar',
	TEXT = 'text',
	TINYTEXT = 'tinytext',
	MEDIUMTEXT = 'mediumtext',
	LONGTEXT = 'longtext',
	DATE = 'date',
	DATETIME = 'datetime',
	TIME = 'time',
	YEAR = 'year',
	BOOL = 'bool',
	BOOLEAN = 'boolean',
	ENUM = 'enum',
	SET = 'set',
	BLOB = 'blob',
	TINYBLOB = 'tinyblob',
	MEDIUMBLOB = 'mediumblob',
	LONGBLOB = 'longblob',
	BINARY = 'binary',
	VARBINARY = 'varbinary',
	JSON = 'json',
}